# api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Dossier API",
    "description" : "The Dossier sync + tracking contract. Implement this spec to build a Dossier-compatible backend the extension and web app can point at. Authenticate at POST /api/authenticate, then send the access token as a Bearer token on every protected request; refresh it at POST /api/refresh.",
    "contact" : { },
    "license" : { },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/api/resumes/{id}" : {
      "get" : {
        "tags" : [ "resume-resource" ],
        "operationId" : "getResume",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeDTO"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "resume-resource" ],
        "operationId" : "updateResume",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResumeDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "resume-resource" ],
        "operationId" : "deleteResume",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "resume-resource" ],
        "operationId" : "partialUpdateResume",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResumeDTO"
              }
            },
            "application/merge-patch+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResumeDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile" : {
      "get" : {
        "tags" : [ "profile-sync" ],
        "summary" : "Get my bio",
        "description" : "The current user's bio payload, or 404 if none saved yet.",
        "operationId" : "getProfile",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BioDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "profile-sync" ],
        "summary" : "Upsert my bio",
        "description" : "Create or overwrite the current user's single bio.",
        "operationId" : "putProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileVM"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BioDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
    "/api/profile/resumes/{id}" : {
      "put" : {
        "tags" : [ "profile-sync" ],
        "summary" : "Update a resume",
        "description" : "Update one of the current user's resumes; 404 if not owned.",
        "operationId" : "updateResume_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResumeDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "profile-sync" ],
        "summary" : "Delete a resume",
        "description" : "Delete one of the current user's resumes (and its stored file); 404 if not owned.",
        "operationId" : "deleteResume_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/resumes" : {
      "get" : {
        "tags" : [ "resume-resource" ],
        "operationId" : "getAllResumes",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Zero-based page index (0..N)",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The size of the page to be returned",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "eagerload",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResumeDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "resume-resource" ],
        "operationId" : "createResume",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResumeDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/{id}/file" : {
      "get" : {
        "tags" : [ "resume-file-resource" ],
        "operationId" : "download",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "resume-file-resource" ],
        "operationId" : "upload",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "resume-file-resource" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/profile/resumes" : {
      "get" : {
        "tags" : [ "profile-sync" ],
        "summary" : "List my resumes",
        "description" : "All resumes owned by the current user.",
        "operationId" : "listResumes",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ResumeDTO"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "profile-sync" ],
        "summary" : "Create a resume",
        "description" : "Create a resume owned by the current user (server fills object key, status, createdAt).",
        "operationId" : "createResume_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResumeDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ResumeDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "label" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "getr2ObjectKey" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "parsedJson" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "NEEDS_REVIEW", "CONFIRMED" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "archived" : {
            "type" : "boolean"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserDTO"
          }
        },
        "required" : [ "createdAt", "getr2ObjectKey", "label", "status" ]
      },
      "UserDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "login" : {
            "type" : "string"
          }
        }
      },
      "ProfileVM" : {
        "type" : "object",
        "properties" : {
          "payload" : {
            "type" : "string"
          }
        },
        "required" : [ "payload" ]
      },
      "BioDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "payload" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserDTO"
          }
        },
        "required" : [ "updatedAt" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "JWT access token issued by POST /api/authenticate (or refreshed via POST /api/refresh).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}